simple calculation

+++ src/chunk.rs
@@ -29,9 +29,9 @@ impl Chunk {
         self.lines.push(line);
     }
 
-    pub fn add_constant(&mut self, value: Value) -> u8 {
+    pub fn add_constant(&mut self, value: Value) -> usize {
         self.constants.write(value);
-        (self.constants.values.len() - 1) as u8
+        self.constants.values.len() - 1
     }
 }
